api -> server に rename, /api を別ファイルに分ける
